+++ app/src/main/java/dev/pandesal/sbp/presentation/transactions/newtransaction/NewTransactionScreen.kt
@@ -1025,8 +1025,7 @@ private fun NewTransactionScreen(
                                     cutoffDays,
                                     reminderEnabled
                                 )
-                            },
-                            enabled = canSave.value,
+                            }
                         ) {
                             Icon(Icons.Default.Check, "Localized description")
                         }